+++ plugins/AtsumaruAutoplayFix.js
@@ -1,7 +1,7 @@
 //=============================================================================
 // AtsumaruAutoplayFix.js
 //
-// Copyright (c) 2018-2019 RPGアツマール開発チーム(https://game.nicovideo.jp/atsumaru)
+// Copyright (c) 2018-2020 RPGアツマール開発チーム(https://game.nicovideo.jp/atsumaru)
 // Released under the MIT license
 // http://opensource.org/licenses/mit-license.php
 //=============================================================================
