+++ plugins/AtsumaruGetSelfInformation.js
@@ -1,7 +1,7 @@
 //=============================================================================
-// AtsumaruGetSelfInformationExperimental.js
+// AtsumaruGetSelfInformation.js
 //
-// Copyright (c) 2018-2019 RPGアツマール開発チーム(https://game.nicovideo.jp/atsumaru)
+// Copyright (c) 2018-2020 RPGアツマール開発チーム(https://game.nicovideo.jp/atsumaru)
 // Released under the MIT license
 // http://opensource.org/licenses/mit-license.php
 //=============================================================================
@@ -132,7 +132,7 @@
     }
 
     /*:
-     * @plugindesc RPGアツマールのプレイヤー本人の情報を取得するAPIのための(Experimental版)プラグインです
+     * @plugindesc RPGアツマールのプレイヤー本人の情報を取得するAPIのためのプラグインです
      * @author RPGアツマール開発チーム
      *
      * @param id
@@ -198,8 +198,8 @@
      *   その時セーブしたセーブデータの状態が不確定になりますので、
      *   可能な限り「並列処理」以外のトリガーでご利用ください。
      */
-    var parameters = toTypedParameters(PluginManager.parameters("AtsumaruGetSelfInformationExperimental"));
-    var getSelfInformation = window.RPGAtsumaru && window.RPGAtsumaru.experimental && window.RPGAtsumaru.experimental.user && window.RPGAtsumaru.experimental.user.getSelfInformation;
+    var parameters = toTypedParameters(PluginManager.parameters("AtsumaruGetSelfInformation"));
+    var getSelfInformation = window.RPGAtsumaru && window.RPGAtsumaru.user.getSelfInformation;
     ensureValidVariableIds(parameters);
     prepareBindPromise();
     addPluginCommand({
