+++ plugins/AtsumaruGlobalSignal.js
@@ -1,7 +1,7 @@
 //=============================================================================
-// AtsumaruGlobalSignalExperimental.js
+// AtsumaruGlobalSignal.js
 //
-// Copyright (c) 2018-2019 RPGアツマール開発チーム(https://game.nicovideo.jp/atsumaru)
+// Copyright (c) 2018-2020 RPGアツマール開発チーム(https://game.nicovideo.jp/atsumaru)
 // Released under the MIT license
 // http://opensource.org/licenses/mit-license.php
 //=============================================================================
@@ -153,7 +153,7 @@
     }
 
     /*:
-     * @plugindesc RPGアツマールのグローバルシグナルのための(Experimental版)プラグインです
+     * @plugindesc RPGアツマールのグローバルシグナルのためのプラグインです
      * @author RPGアツマール開発チーム
      *
      * @param signalData
@@ -219,8 +219,8 @@
      *   その時セーブしたセーブデータの状態が不確定になりますので、
      *   可能な限り「並列処理」以外のトリガーでご利用ください。
      */
-    var parameters = toTypedParameters(PluginManager.parameters("AtsumaruGlobalSignalExperimental"));
-    var signal = window.RPGAtsumaru && window.RPGAtsumaru.experimental && window.RPGAtsumaru.experimental.signal;
+    var parameters = toTypedParameters(PluginManager.parameters("AtsumaruGlobalSignal"));
+    var signal = window.RPGAtsumaru && window.RPGAtsumaru.signal;
     var sendGlobalSignal = signal && signal.sendSignalToGlobal;
     var getGlobalSignals = signal && signal.getGlobalSignals;
     ensureValidVariableIds(parameters);
