+++ src/AtsumaruUserSignal.ts
@@ -1,29 +1,29 @@
 /*:
- * @plugindesc RPGアツマールのグローバルシグナルのための(Experimental版)プラグインです
+ * @plugindesc RPGアツマールのユーザーシグナルのためのプラグインです
  * @author RPGアツマール開発チーム
  *
  * @param signalData
  * @type variable
  * @text シグナルデータ
- * @desc グローバルシグナルの取得時に、シグナルデータを代入する変数の番号を指定します。
+ * @desc ユーザーシグナルの取得時に、シグナルデータを代入する変数の番号を指定します。
  * @default 0
  *
  * @param senderId
  * @type variable
  * @text 送信者のユーザーID
- * @desc グローバルシグナルの取得時に、送信者のユーザーIDを代入する変数の番号を指定します。
+ * @desc ユーザーシグナルの取得時に、送信者のユーザーIDを代入する変数の番号を指定します。
  * @default 0
  *
  * @param senderName
  * @type variable
  * @text 送信者のユーザー名
- * @desc グローバルシグナルの取得時に、送信者のユーザー名を代入する変数の番号を指定します。
+ * @desc ユーザーシグナルの取得時に、送信者のユーザー名を代入する変数の番号を指定します。
  * @default 0
  *
  * @param restCount
  * @type variable
  * @text 残シグナル数
- * @desc グローバルシグナルの取得時に、今取得したものを含めた残りのシグナル数を代入する変数の番号を指定します。
+ * @desc ユーザーシグナルの取得時に、今取得したものを含めた残りのシグナル数を代入する変数の番号を指定します。
  * @default 0
  *
  * @param errorMessage
@@ -33,32 +33,32 @@
  * @default 0
  *
  * @help
- * このプラグインは、アツマールAPIの「グローバルシグナル」を利用するためのプラグインです。
+ * このプラグインは、アツマールAPIの「ユーザーシグナル」を利用するためのプラグインです。
  * 詳しくはアツマールAPIリファレンス(https://atsumaru.github.io/api-references/signal)を参照してください。
  *
  * プラグインコマンド（英語版と日本語版のコマンドがありますが、どちらも同じ動作です）:
- *   SendGlobalSignal <signalDataVariableId>
- *   グローバルシグナル送信 <signalDataVariableId>
+ *   SendUserSignal <signalDataVariableId> <userIdVariableId>
+ *   ユーザーシグナル送信 <signalDataVariableId> <userIdVariableId>
  *     # 変数<signalDataVariableId>からシグナルデータを読み取り、
- *          それをグローバルシグナルとして送信します。
- *     # 例: SendGlobalSignal 1
- *     #   : グローバルシグナル送信 1
+ *          それを変数<userIdVariableId>から読み取ったユーザーIDの相手に送信します。
+ *     # 例: SendUserSignal 1 2
+ *     #   : ユーザーシグナル送信 1 2
  *
- *   GetGlobalSignal
- *   グローバルシグナル取得
- *      # まだ取得したことがないグローバルシグナルの中で最も古い一件を読み込み、
+ *   GetUserSignal
+ *   ユーザーシグナル取得
+ *      # まだ取得したことがないユーザーシグナルの中で最も古い一件を読み込み、
  *          プラグインパラメータで指定した変数に値をセットします。
  *      # 残シグナル数が0だった時は、シグナルデータと送信者のユーザーID/名前には0がセットされます。
  *      # 残シグナル数が0か1だった時は、次の取得コマンドで新たなシグナルの受信を試みますので
  *          時間がかかることがあります。スムーズに実行したい場合は、
  *          次の取得コマンドの実行（受信）までに１０秒以上の時間を空けてください。
- *      # 例: GetGlobalSignal
- *      #   : グローバルシグナル取得
+ *      # 例: GetUserSignal
+ *      #   : ユーザーシグナル取得
  *
  * アツマール外（テストプレイや他のサイト、ダウンロード版）での挙動:
- *      SendGlobalSignal（グローバルシグナル送信）
+ *      SendUserSignal（ユーザーシグナル送信）
  *          無視される（エラーメッセージにも何も代入されない）
- *      GetGlobalSignal（グローバルシグナル取得）
+ *      GetUserSignal（ユーザーシグナル取得）
  *          無視される（エラーメッセージにも何も代入されない）
  *
  * ※「並列処理」の中でプラグインコマンドを利用しますと
@@ -67,9 +67,9 @@
  */
 
 import { isNumber } from "./utils/typecheck";
-import { toValidVariableId, toTypedParameters, ensureValidVariableIds } from "./utils/parameter";
+import { toNatural, toValidVariableId, toTypedParameters, ensureValidVariableIds } from "./utils/parameter";
 import { hookStatic, addPluginCommand, prepareBindPromise } from "./utils/rmmvbridge";
-import { GlobalSignal } from "@atsumaru/api-types";
+import { UserSignal } from "@atsumaru/api-types";
 
 interface Parameters {
     signalData: number
@@ -79,25 +79,25 @@ interface Parameters {
     errorMessage: number
 }
 
-interface GlobalSignalStore {
-    signals: GlobalSignal[]
-    lastPoppedSignals: GlobalSignal[]
+interface UserSignalStore {
+    signals: UserSignal[]
+    lastPoppedSignals: UserSignal[]
 }
 
 declare const window: Window;
-const parameters = toTypedParameters(PluginManager.parameters("AtsumaruGlobalSignalExperimental")) as Parameters;
-const signal = window.RPGAtsumaru && window.RPGAtsumaru.experimental && window.RPGAtsumaru.experimental.signal;
-const sendGlobalSignal = signal && signal.sendSignalToGlobal;
-const getGlobalSignals = signal && signal.getGlobalSignals;
+const parameters = toTypedParameters(PluginManager.parameters("AtsumaruUserSignal")) as Parameters;
+const signal = window.RPGAtsumaru && window.RPGAtsumaru.signal;
+const sendUserSignal = signal && signal.sendSignalToUser;
+const getUserSignals = signal && signal.getUserSignals;
 
 ensureValidVariableIds(parameters);
 prepareBindPromise();
 
 addPluginCommand({
-    SendGlobalSignal,
-    "グローバルシグナル送信": SendGlobalSignal,
-    GetGlobalSignal,
-    "グローバルシグナル取得": GetGlobalSignal
+    SendUserSignal,
+    "ユーザーシグナル送信": SendUserSignal,
+    GetUserSignal,
+    "ユーザーシグナル取得": GetUserSignal
 });
 
 hookStatic(DataManager, "createGameObjects", origin => function(this: typeof DataManager) {
@@ -111,30 +111,30 @@ hookStatic(DataManager, "extractSaveContents", origin => function(this: typeof D
 });
 
 function initialFetch() {
-    if (!$gameSystem._globalSignalStoreForRPGAtsumaruPlugin && getGlobalSignals) {
-        $gameSystem._globalSignalStoreForRPGAtsumaruPlugin = { signals: [], lastPoppedSignals: [] };
-        fetchGlobalSignal(getGlobalSignals, $gameSystem._globalSignalStoreForRPGAtsumaruPlugin);
+    if (!$gameSystem._userSignalStoreForRPGAtsumaruPlugin && getUserSignals) {
+        $gameSystem._userSignalStoreForRPGAtsumaruPlugin = { signals: [], lastPoppedSignals: [] };
+        fetchUserSignal(getUserSignals, $gameSystem._userSignalStoreForRPGAtsumaruPlugin);
     }
 }
 
-function fetchGlobalSignal(getGlobalSignals: () => Promise<GlobalSignal[]>, store: GlobalSignalStore) {
-    return getGlobalSignals().then(
-        globalSignals => {
-            globalSignals.sort((a, b) => b.createdAt - a.createdAt);
+function fetchUserSignal(getUserSignals: () => Promise<UserSignal[]>, store: UserSignalStore) {
+    return getUserSignals().then(
+        userSignals => {
+            userSignals.sort((a, b) => b.createdAt - a.createdAt);
             if (store.lastPoppedSignals.length > 0) {
-                const lastIndex = globalSignals.map(signal => signal.createdAt)
+                const lastIndex = userSignals.map(signal => signal.createdAt)
                     .lastIndexOf(store.lastPoppedSignals[0].createdAt);
                 if (lastIndex !== -1) {
-                    globalSignals = globalSignals.slice(0, lastIndex + 1);
-                    difference(globalSignals, store.lastPoppedSignals);
+                    userSignals = userSignals.slice(0, lastIndex + 1);
+                    difference(userSignals, store.lastPoppedSignals);
                 }
             }
-            store.signals = globalSignals;
+            store.signals = userSignals;
         }
     );
 }
 
-function difference(signals: GlobalSignal[], excludes: GlobalSignal[]) {
+function difference(signals: UserSignal[], excludes: UserSignal[]) {
     const excludeIds = excludes.map(signal => signal.id);
     const excludeCreatedAt = excludes[0].createdAt;
     for (let index = signals.length - 1; index >= 0 && signals[index].createdAt === excludeCreatedAt; index--) {
@@ -144,28 +144,30 @@ function difference(signals: GlobalSignal[], excludes: GlobalSignal[]) {
     }
 }
 
-function SendGlobalSignal(this: Game_Interpreter, command: string, signalDataVariableIdStr?: string) {
+function SendUserSignal(this: Game_Interpreter, command: string, signalDataVariableIdStr?: string, userIdVariableIdStr?: string) {
     const signalDataVariableId = toValidVariableId(signalDataVariableIdStr, command, "signalDataVariableId");
     const signalData = String($gameVariables.value(signalDataVariableId));
-    if (sendGlobalSignal) {
-        this.bindPromiseForRPGAtsumaruPlugin(sendGlobalSignal(signalData),
+    const userIdVariableId = toValidVariableId(userIdVariableIdStr, command, "userIdVariableId");
+    const userId = toNatural($gameVariables.value(userIdVariableId), command, "userId");
+    if (sendUserSignal) {
+        this.bindPromiseForRPGAtsumaruPlugin(sendUserSignal(userId, signalData),
             () => $gameVariables.setValue(parameters.errorMessage, 0),
             error => $gameVariables.setValue(parameters.errorMessage, error.message)
         );
     }
 }
 
-function GetGlobalSignal(this: Game_Interpreter) {
-    if (getGlobalSignals) {
-        const store = $gameSystem._globalSignalStoreForRPGAtsumaruPlugin;
+function GetUserSignal(this: Game_Interpreter) {
+    if (getUserSignals) {
+        const store = $gameSystem._userSignalStoreForRPGAtsumaruPlugin;
         if (store.signals.length === 0) {
             this.bindPromiseForRPGAtsumaruPlugin(
-                throttlePromise(() => fetchGlobalSignal(getGlobalSignals, store)),
-                () => setGlobalSignal(store),
+                throttlePromise(() => fetchUserSignal(getUserSignals, store)),
+                () => setUserSignal(store),
                 error => $gameVariables.setValue(parameters.errorMessage, error.message)
             );
         } else {
-            setGlobalSignal(store);
+            setUserSignal(store);
         }
     }
 }
@@ -185,13 +187,13 @@ function throttlePromise<T>(promise: () => Promise<T>): Promise<T> {
     });
 }
 
-function setGlobalSignal(store: GlobalSignalStore) {
+function setUserSignal(store: UserSignalStore) {
     $gameVariables.setValue(parameters.signalData, 0);
     $gameVariables.setValue(parameters.senderId, 0);
     $gameVariables.setValue(parameters.senderName, 0);
     $gameVariables.setValue(parameters.restCount, store.signals.length);
     $gameVariables.setValue(parameters.errorMessage, 0);
-    const signal: GlobalSignal | undefined = store.signals.pop();
+    const signal: UserSignal | undefined = store.signals.pop();
     if (signal) {
         $gameVariables.setValue(parameters.signalData, isNumber(signal.data) ? +signal.data : signal.data);
         $gameVariables.setValue(parameters.senderId, signal.senderId);
